Changed size_t to int.

+++ src/teal.h
@@ -1,10 +1,8 @@
 #ifndef _TEAL_H_
 #define _TEAL_H_
 
-#include <stddef.h>
-
 // util.c
-void read_stdin(char *buf, size_t count);
+void read_stdin(char *buf, int count);
 
 // lex.c
 #define OP_ADD 0
