+++ src/util.c
@@ -2,8 +2,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void read_stdin(char *buf, size_t count) {
-  size_t rc = fread(buf, 1, count, stdin);
+void read_stdin(char *buf, int count) {
+  int rc = fread(buf, 1, count, stdin);
   buf[rc] = '\0';
 
   if (getc(stdin) != EOF) {
